+++ SofaKernel/modules/SofaBaseTopology/EdgeSetTopologyContainer.h
@@ -128,12 +128,12 @@ public:
      * The difference to getNbEdges() is that this method does not generate the edge array if it does not exist.
      * @return the number of edges.
      */
-    unsigned int getNumberOfEdges() const;
+    size_t getNumberOfEdges() const;
 
     /** \brief Returns the number of topological element of the current topology.
      * This function avoids to know which topological container is in used.
      */
-    virtual unsigned int getNumberOfElements() const override;
+    virtual size_t getNumberOfElements() const override;
 
 
     /** \brief Returns the number of connected components from the graph containing all edges and give, for each vertex, which component it belongs to  (use BOOST GRAPH LIBRAIRY)
@@ -176,7 +176,7 @@ public:
     virtual bool checkConnexity() override;
 
     /// Returns the number of connected component.
-    virtual unsigned int getNumberOfConnectedComponent() override;
+    virtual size_t getNumberOfConnectedComponent() override;
 
     /// Returns the set of element indices connected to an input one (i.e. which can be reached by topological links)
     virtual const VecEdgeID getConnectedElement(EdgeID elem) override;
